+++ extensions/cornerstone/src/services/SegmentationService/SegmentationServiceTypes.ts
@@ -20,6 +20,8 @@ type Segment = {
   isVisible: boolean;
   // whether the segment is locked
   isLocked: boolean;
+  // display texts
+  displayText?: string[];
 };
 
 type Segmentation = {