+++ modes/tmtv/src/index.js
@@ -44,8 +44,12 @@ function modeFactory({ modeConfiguration }) {
      * Lifecycle hooks
      */
     onModeEnter: ({ servicesManager, extensionManager, commandsManager }) => {
-      const { toolbarService, toolGroupService, hangingProtocolService, displaySetService } =
-        servicesManager.services;
+      const {
+        toolbarService,
+        toolGroupService,
+        hangingProtocolService,
+        displaySetService,
+      } = servicesManager.services;
 
       const utilityModule = extensionManager.getModuleEntry(
         '@ohif/extension-cornerstone.utilityModule.tools'
@@ -89,8 +93,8 @@ function modeFactory({ modeConfiguration }) {
         'Crosshairs',
         'Pan',
         'SyncToggle',
-        'RectangleROIStartEndThreshold',
       ]);
+      toolbarService.createButtonSection('tmtvToolbox', ['RectangleROIStartEndThreshold']);
 
       // For the hanging protocol we need to decide on the window level
       // based on whether the SUV is corrected or not, hence we can't hard