+++ platform/ui/src/components/index.js
@@ -29,7 +29,7 @@ import NavBar from './NavBar';
 import Notification from './Notification';
 import Select from './Select';
 import SegmentationTable from './SegmentationTable';
-import SegmentationGroupTable from './SegmentationGroupTable';
+import { SegmentationGroupTable, SegmentationGroupTableExpanded } from './SegmentationGroupTable';
 import SidePanel from './SidePanel';
 import SplitButton from './SplitButton';
 import StudyBrowser from './StudyBrowser';
@@ -80,6 +80,7 @@ import PanelSection from './PanelSection';
 import AdvancedToolbox from './AdvancedToolbox';
 import InputDoubleRange from './InputDoubleRange';
 import LegacyButtonGroup from './LegacyButtonGroup';
+import LabellingFlow from './Labelling';
 import SwitchButton, { SwitchLabelLocation } from './SwitchButton';
 import * as AllInOneMenu from './AllInOneMenu';
 import ViewportActionArrows from './ViewportActionArrows';
@@ -89,7 +90,6 @@ import { ToolSettings } from './AdvancedToolbox';
 import { Toolbox } from './Toolbox';
 import InvestigationalUseDialog from './InvestigationalUseDialog';
 import MeasurementItem from './MeasurementTable/MeasurementItem';
-
 import LayoutPreset from './LayoutPreset';
 
 export {
@@ -144,6 +144,7 @@ export {
   Select,
   SegmentationTable,
   SegmentationGroupTable,
+  SegmentationGroupTableExpanded,
   SidePanel,
   SplitButton,
   StudyBrowser,
@@ -181,6 +182,7 @@ export {
   ViewportPane,
   ViewportOverlay,
   WindowLevelMenuItem,
+  LabellingFlow,
   HeaderPatientInfo,
   LayoutPreset,
   LegacySplitButton,